requirements.txt
google-cloud-aiplatform
google-adk
requests  
PyPDF2
python-dotenv